requirements.txt
Django>=3.2
djangorestframework
python-dotenv
